万恶的IE

+++ 01SOURCE/base-common-web/src/main/java/cn/doublepoint/common/port/adapter/template/persistence/sys/admin/AdminService.java
@@ -26,7 +26,7 @@ public interface AdminService {
 	 * @param id
 	 * @return
 	 */
-	public SysAdmin getById(long id);
+	public SysAdmin getById(Integer id);
 	
 	/**
 	 * 移除
